+++ noodles-cram/examples/cram_query.rs
@@ -39,7 +39,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let query = reader.query(&header, &region)?;
 
     for result in query {
-        let record = result.and_then(|record| record.try_into_alignment_record(&header))?;
+        let record = result?;
         writer.write_alignment_record(&header, &record)?;
     }