+++ noodles-cram/src/record/convert.rs
@@ -4,13 +4,18 @@ use noodles_sam as sam;
 
 use super::{features::Cigar, Record};
 
-impl Record {
+impl Record<'_> {
     /// Converts this CRAM record to an alignment record.
     pub fn try_into_alignment_record(
         self,
-        header: &sam::Header,
+        _header: &sam::Header,
     ) -> io::Result<sam::alignment::RecordBuf> {
-        let mut builder = sam::alignment::RecordBuf::builder();
+        let data = sam::alignment::Record::data(&self)
+            .iter()
+            .map(|result| result.and_then(|(tag, value)| value.try_into().map(|v| (tag, v))))
+            .collect::<io::Result<_>>()?;
+
+        let mut builder = sam::alignment::RecordBuf::builder().set_data(data);
 
         if let Some(read_name) = self.name {
             builder = builder.set_name(read_name);
@@ -48,29 +53,6 @@ impl Record {
             .set_sequence(self.sequence)
             .set_quality_scores(self.quality_scores);
 
-        let mut data = self.data;
-        maybe_insert_read_group(&mut data, header.read_groups(), self.read_group_id)?;
-        builder = builder.set_data(data);
-
         Ok(builder.build())
     }
-}
-
-fn maybe_insert_read_group(
-    data: &mut sam::alignment::record_buf::Data,
-    read_groups: &sam::header::ReadGroups,
-    read_group_id: Option<usize>,
-) -> io::Result<()> {
-    use sam::alignment::{record::data::field::Tag, record_buf::data::field::Value};
-
-    if let Some(id) = read_group_id {
-        let name = read_groups
-            .get_index(id)
-            .map(|(name, _)| name)
-            .ok_or_else(|| io::Error::new(io::ErrorKind::InvalidInput, "invalid read group ID"))?;
-
-        data.insert(Tag::READ_GROUP, Value::String(name.clone()));
-    }
-
-    Ok(())
 }